faq_multilingual/requirements.txt
# Django and REST framework
Django==5.0.1
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# WYSIWYG Editor
django-ckeditor==6.7.0
django-js-asset==2.2.0

# Translation
googletrans==3.1.0a0

# Caching
redis==5.0.1
django-redis==5.4.0

# CORS and Security
django-cors-headers==4.3.1
django-environ==0.11.2

# Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Documentation
drf-yasg==1.21.7

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.20.0